add error handling and size slider

+++ main.py
@@ -14,6 +14,8 @@ class Watermarker(customtkinter.CTk):
         self.columnconfigure((0, 1, 2, 3), weight=1, uniform='a')
         self.rowconfigure(0, weight=1)
 
+        self.watermark_size = 20
+
         self.my_image = Image.open('default-placeholder.png')
         self.image_ratio = self.my_image.size[0] / self.my_image.size[1]
         self.resized_image_tk = None
@@ -25,6 +27,7 @@ class Watermarker(customtkinter.CTk):
         self.path_field_label = customtkinter.CTkLabel(self.widget_frame, text='File Path: ', font=('verdana', 12))
         self.path_field_label.grid(row=0, column=0, sticky='w', padx=15)
 
+        # file path entry field:
         self.path_field = customtkinter.CTkEntry(self.widget_frame)
         self.path_field.grid(row=1, column=0, padx=10, columnspan=2, sticky='nsew')
 
@@ -36,14 +39,18 @@ class Watermarker(customtkinter.CTk):
                                                            font=('verdana', 12), anchor='w')
         self.watermark_text_label.grid(row=3, column=0, sticky='w', padx=15)
 
+        # watermark text entry field:
         self.watermark_text = customtkinter.CTkEntry(self.widget_frame)
         self.watermark_text.grid(row=4, column=0, padx=10, pady=(0, 10), columnspan=2, sticky='nsew')
 
+        self.slider = customtkinter.CTkSlider(self.widget_frame, from_=0, to=100, command=self.set_watermark_size)
+        self.slider.grid(row=5, column=0, columnspan=2, sticky='nsew', pady=(0,10))
+
         self.add_button = customtkinter.CTkButton(self.widget_frame, text='Add Watermark', command=self.add_watermark)
-        self.add_button.grid(row=5, column=0, padx=10, sticky='w')
+        self.add_button.grid(row=6, column=0, padx=10, sticky='w')
 
         self.save_button = customtkinter.CTkButton(self.widget_frame, text='Save', command=self.save_file)
-        self.save_button.grid(row=5, column=1, padx=10, sticky='e')
+        self.save_button.grid(row=6, column=1, padx=10, sticky='e')
 
         self.widget_frame.grid(row=0, column=0, sticky='nsew')
 
@@ -67,20 +74,29 @@ class Watermarker(customtkinter.CTk):
         self.canvas.create_image(int(event.width / 2), int(event.height / 2), anchor='center',
                                  image=self.resized_image_tk)
 
+    def set_watermark_size(self, size):
+        if size > 1 and self.watermark_text.get():
+            self.watermark_size = size
+            self.add_watermark()
+
     def load_file_name(self):
         filename = filedialog.askopenfilename()
-        self.path_field.delete(0, END)  # Clear previous text
-        self.path_field.insert(END, filename)
+        if filename != '':
+            self.path_field.delete(0, END)  # Clear previous text
+            self.path_field.insert(END, filename)
 
-        self.open_files()
+            self.open_files()
+
 
     def open_files(self):
-        filename = self.path_field.get()
-        self.my_image = Image.open(filename)
-        self.image_ratio = self.my_image.size[0] / self.my_image.size[1]
+        if self.path_field.get():
+            filename = self.path_field.get()
+            self.my_image = Image.open(filename)
+            self.image_ratio = self.my_image.size[0] / self.my_image.size[1]
 
-        # Trigger the <Configure> event with the canvas dimensions
-        self.canvas.event_generate('<Configure>', width=self.canvas.winfo_width(), height=self.canvas.winfo_height())
+            # Trigger the <Configure> event with the canvas dimensions
+            self.canvas.event_generate('<Configure>', width=self.canvas.winfo_width(), height=self.canvas.winfo_height())
+
 
     def save_file(self):
         # Use asksaveasfile to get the file path and name
@@ -99,14 +115,14 @@ class Watermarker(customtkinter.CTk):
         watermarked_image = self.my_image.copy()
 
         draw = ImageDraw.Draw(watermarked_image)
-        font = ImageFont.truetype('verdana', 20)
+        font = ImageFont.truetype('verdana', self.watermark_size)
 
         # Get the text size
         textbbox = draw.textbbox((0, 0), f'© {watermark_text}', font=font)
         # Extract width and height from the textbbox tuple
         text_width, text_height = textbbox[2] - textbbox[0], textbbox[3] - textbbox[1]
 
-        text_position = (watermarked_image.width - text_width - 5, watermarked_image.height - text_height - 10)
+        text_position = (watermarked_image.width - text_width - 35, watermarked_image.height - text_height - 35)
         draw.text(text_position, f'© {watermark_text}', fill=(255, 255, 255, 200), font=font)
 
         self.my_image = watermarked_image
